+++ test/unit/gmrs/io/csv_test.clj
@@ -3,20 +3,21 @@
             [clojure.java.io :as io]
             [gmrs.io.csv :refer :all]))
 
+(def expected-example
+  [{:name "Felipe" :city "Madrid" :age "20"}
+   {:name "Maria" :city "Aguilas" :age "30"}
+   {:name "Ernesto" :city "La Coruña" :age "25"}])
+
 (deftest test-csv-give
   (with-open [reader (io/reader "test/_res/example.csv")]
-    (let [give (csv-give reader true)]
-      (is (= [{:name "Felipe" :city "Madrid" :age "20"}
-              {:name "Maria" :city "Aguilas" :age "30"}
-              {:name "Ernesto" :city "La Coruña" :age "25"}]
-             (give {:page-size 10}))
+    (let [give ((csv-give reader true) {:page-size 10})]
+      (is (= expected-example
+             (first give))
           "reading contents from a CSV file")))
   (with-open [reader (io/reader "test/_res/example.csv")]
-    (let [give (csv-give reader)]
-      (is (= [{:name "Felipe" :city "Madrid" :age "20"}
-              {:name "Maria" :city "Aguilas" :age "30"}
-              {:name "Ernesto" :city "La Coruña" :age "25"}]
-             (give {:page-size 10}))
+    (let [give ((csv-give reader true) {:page-size 10})]
+      (is (= expected-example
+             (first give))
           "default wrap? arg value"))))
 
 ; (run-tests 'gmrs.io.csv-test)